fix: Move package under sink/type/tablesink
Add new configuration option for file import send timeout.
