+++ internal/pkg/service/stream/config/config.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/distribution"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/etcdclient"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/keboolasink/bridge"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/sink"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/source"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage"
@@ -32,7 +31,6 @@ type Config struct {
 	Source          source.Config       `configKey:"source"`
 	Sink            sink.Config         `configKey:"sink"`
 	Storage         storage.Config      `configKey:"storage"`
-	KeboolaBridge   bridge.Config       `configKey:"keboolaBridge"`
 }
 
 type Patch struct {
@@ -61,7 +59,6 @@ func New() Config {
 		Source:          source.NewConfig(),
 		Sink:            sink.NewConfig(),
 		Storage:         storage.NewConfig(),
-		KeboolaBridge:   bridge.NewConfig(),
 	}
 }
