+++ internal/pkg/service/stream/sink/type/tablesink/config.go
@@ -1,9 +1,15 @@
 package tablesink
 
-type Config struct{}
+import "github.com/keboola/keboola-as-code/internal/pkg/service/stream/sink/type/tablesink/keboola"
+
+type Config struct {
+	Keboola keboola.Config `configKey:"keboola"`
+}
 
 type ConfigPatch struct{}
 
 func NewConfig() Config {
-	return Config{}
+	return Config{
+		Keboola: keboola.NewConfig(),
+	}
 }
