+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/bucket.go
