+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/fixtures/create_sink_snapshot.txt
