+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/fixtures/delete_files_ops.txt
