+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/fixtures/disable_sink_snapshot.txt
