+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/fixtures/enable_sink_ops.txt
