+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/schema/file_test.go
