+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/schema/schema.go
