+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/schema/token_test.go
