+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/slice.go
