+++ internal/pkg/service/stream/sink/type/tablesink/keboola/bridge/test/table.go
